+++ packages/scenes/src/core/SceneTimeRange.tsx
@@ -1,11 +1,13 @@
-import { dateMath, getTimeZone, TimeRange, toUtc } from '@grafana/data';
+import { getTimeZone, TimeRange } from '@grafana/data';
 import { TimeZone } from '@grafana/schema';
 
 import { SceneObjectUrlSyncConfig } from '../services/SceneObjectUrlSyncConfig';
 
 import { SceneObjectBase } from './SceneObjectBase';
-import { SceneTimeRangeLike, SceneTimeRangeState, SceneObjectUrlValues, SceneObjectUrlValue } from './types';
+import { SceneTimeRangeLike, SceneTimeRangeState, SceneObjectUrlValues } from './types';
 import { getClosest } from './sceneGraph/utils';
+import { parseUrlParam } from '../utils/parseUrlParam';
+import { evaluateTimeRange } from '../utils/evaluateTimeRange';
 
 export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> implements SceneTimeRangeLike {
   protected _urlSync = new SceneObjectUrlSyncConfig(this, { keys: ['from', 'to'] });
@@ -132,52 +134,4 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
     update.value = evaluateTimeRange(update.from ?? this.state.from, update.to ?? this.state.to, this.getTimeZone());
     this.setState(update);
   }
-}
-
-export function parseUrlParam(value: SceneObjectUrlValue): string | null {
-  if (typeof value !== 'string') {
-    return null;
-  }
-
-  if (value.indexOf('now') !== -1) {
-    return value;
-  }
-
-  if (value.length === 8) {
-    const utcValue = toUtc(value, 'YYYYMMDD');
-    if (utcValue.isValid()) {
-      return utcValue.toISOString();
-    }
-  } else if (value.length === 15) {
-    const utcValue = toUtc(value, 'YYYYMMDDTHHmmss');
-    if (utcValue.isValid()) {
-      return utcValue.toISOString();
-    }
-  } else if (value.length === 24) {
-    const utcValue = toUtc(value);
-    return utcValue.toISOString();
-  }
-
-  const epoch = parseInt(value, 10);
-  if (!isNaN(epoch)) {
-    return toUtc(epoch).toISOString();
-  }
-
-  return null;
-}
-
-export function evaluateTimeRange(
-  from: string,
-  to: string,
-  timeZone: TimeZone,
-  fiscalYearStartMonth?: number
-): TimeRange {
-  return {
-    from: dateMath.parse(from, false, timeZone, fiscalYearStartMonth)!,
-    to: dateMath.parse(to, true, timeZone, fiscalYearStartMonth)!,
-    raw: {
-      from: from,
-      to: to,
-    },
-  };
 }
